Build system: compile QML for Qt 5.11 onwards
Compile the QML code from Qt 5.11 onwards. This is a very rudimentary
first attempt. Do not call qtquick_compiler_add_resources() more then
once. This results in CMake Error: Attempt to add a custom rule to
output ...  which already has a custom rule. This is the reason a list
of resources is constructed, and passed to qtquick_compiler_add_resources()
just once.

This all seems to work on Linux, building desktop, desktop-on-mobile, and
a native Android build. Interestingly, there seems no significant
performance gain on startup of the app on device (Android).
Tested on 2 identical devices opening the app simultaneous the same
(big) logbook.

Notice that Qt advises to use the QML compiler only for release
builds. No debugging is given as the main reason for this. This
commit does only look at Qt version.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>